+++ R/todo.R
@@ -43,6 +43,12 @@ do_todo <- function(text, pattern = NULL, path = path, ...) {
   }
 
   files <- if (is_dir(path)) {
+
+    if (!length(list.files(path, pattern = "\\.Rproj$"))) {
+      message("Did not search for TODOS in ", norm_path(path))
+      return(invisible(NULL))
+    }
+
     list.files(
       path,
       pattern = "\\.r(md)?$",